- Correct user id retrieval in req - authController.js
Move fetchUsertrips() from authController.js into tripController.js

Revise routes as needed and validate a proper endpoint can be reached through trip/. Now can be reached by trip/readAll
